docs: begin separation of dataset and feature py files
